mcp_server/github_pat_mcp/requirements.txt
fastmcp>=3.0.0b
httpx>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1